requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-mysql==5.2.0
apache-airflow-providers-google==10.8.0
pandas>=2.0.0
pandas-gbq==0.23.1
PyMySQL>=1.0.2
google-cloud-bigquery>=3.10.0
python-dotenv>=1.0.0
pydantic==1.10.15
sqlalchemy>=2.0.0